Fix memory leaks, HNSW corruption, and search latency
- Optimize WAL memory footprint during Flush with GetMaxSeqID
- Resolve HNSW spatial corruption when resurrecting soft-deleted nodes
- Cap searchK exponential fallback to prevent brute-force O(N) degradation
- Eliminate SysDB vector metadata eager-loading memory bloat

+++ internal/index/hnsw/snapshot.go
@@ -228,7 +228,7 @@ func LoadHNSWFromSnapshot(path string) (*HNSWIndex, uint64, error) {
 		nodes:          nodes,
 		idToNode:       idToNode,
 		rng:            rand.New(rand.NewSource(time.Now().UnixNano())),
-		deleted:        make(map[uint64]bool),
+		deleted:        make(map[int]bool),
 	}
 
 	return h, hdr.SeqID, nil
